requirements.txt
Django==4.0.1
django-environ==0.9.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
django-cors-headers==3.13.0
social-auth-app-django==5.0.0
django-filter==22.1
requests==2.28.1
marshmallow==3.19.0
marshmallow-dataclass==8.5.10
pytest-django==4.5.2
factory-boy==3.2.1
pytest-factoryboy==2.5.0
